refactor: style 파일 정리
