+++ src/features/post/components/StepFunnel/Step1/style.css.ts
@@ -1,0 +1,16 @@
+import { css } from '@styled-system/css';
+
+export const Wrapper = css({
+  display: 'flex',
+  flexDirection: 'column',
+  height: '100%',
+});
+
+export const Container = css({
+  display: 'flex',
+  gap: '0.58rem',
+  height: '100%',
+  position: 'relative',
+  justifyContent: 'center',
+  alignItems: 'center',
+});
