+++ src/features/post/components/StepFunnel/Step2/index.tsx
@@ -1,5 +1,6 @@
 import Token from '@/common/components/Token';
 import * as s from './style.css';
+import * as c from '../style.css';
 
 import TagOptionBtn from '@/common/components/TagOptionBtn';
 import { PRODUCT_TYPES_MAP, type ProductType } from '@/libs/types/item';
@@ -14,11 +15,11 @@ const Step2 = () => {
 
   return (
     <div>
-      <header className={s.Head}>
+      <header className={c.Head}>
         제품 타입을 선택해 주세요
         <Token>복수 선택 가능</Token>
       </header>
-      <div className={s.Content}>
+      <div className={c.Content}>
         <div className={s.Grid}>
           {(Object.keys(PRODUCT_TYPES_MAP) as ProductType[]).map(type => (
             <TagOptionBtn
