+++ src/features/post/components/StepFunnel/Step2/style.css.ts
@@ -1,0 +1,7 @@
+import { css } from '@styled-system/css';
+
+export const Grid = css({
+  display: 'grid',
+  gridTemplateColumns: 'repeat(2, 1fr)',
+  gap: '0.62rem',
+});
