+++ src/features/post/components/StepFunnel/Step3/index.tsx
@@ -1,0 +1,45 @@
+import Chip from '@/common/components/Chip';
+import * as c from '../style.css';
+import { TRADE_TYPES_MAP, type TradeType } from '@/libs/types/post';
+import Token from '@/common/components/Token';
+import InputField from '../../InputField';
+import { useState } from 'react';
+
+const Step3 = () => {
+  const [selectedTypes, setSelectedTypes] = useState<TradeType[]>([]);
+
+  const handleSelectType = (type: TradeType) => {
+    setSelectedTypes(prev => (prev.includes(type) ? prev.filter(t => t !== type) : [...prev, type]));
+  };
+
+  return (
+    <div>
+      <header className={c.Head}>거래 방식을 선택해 주세요</header>
+      <div className={c.Content}>
+        <div className={c.DetailContent}>
+          <div className={c.ContentHead}>
+            거래 방식을 선택해 주세요
+            <Token>복수 선택 가능</Token>
+          </div>
+          <div className={c.ChipContainer}>
+            {(Object.entries(TRADE_TYPES_MAP) as [TradeType, string][])
+              .filter(([key]) => key !== 'DIRECT_AND_PARCEL')
+              .map(([key, label]) => (
+                <Chip key={key} isSelected={selectedTypes.includes(key)} onClick={() => handleSelectType(key)}>
+                  {label}
+                </Chip>
+              ))}
+          </div>
+        </div>
+        <div className={c.DetailContent}>
+          <div className={c.DetailContent}>
+            <span>직거래 장소를 입력해 주세요</span>
+            <InputField />
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Step3;
