+++ src/features/post/components/StepFunnel/Step5/style.css.ts
@@ -1,0 +1,17 @@
+import { css } from '@styled-system/css';
+
+export const ProductDesc = css({
+  // step5
+  display: 'flex',
+  flexDirection: 'column',
+  width: '100%',
+  gap: '1.25rem',
+});
+
+export const SelectPhotoContainer = css({
+  // step5
+  display: 'flex',
+  gap: '0.625rem',
+  alignItems: 'center',
+  flexWrap: 'wrap',
+});
