+++ src/features/post/components/StepFunnel/Step6/style.css.ts
@@ -1,0 +1,58 @@
+import { css } from '@styled-system/css';
+
+export const PriceTypeContent = css({
+  // step6
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '0.75rem',
+});
+
+export const PriceInputField = css({
+  // step6
+  display: 'flex',
+  alignItems: 'center',
+  gap: '0.75rem',
+  alignSelf: 'stretch',
+  color: '80',
+});
+
+export const PriceDesc = css({
+  // step6
+  display: 'flex',
+  justifyContent: 'start',
+  alignItems: 'start',
+  flexDirection: 'column',
+  width: '100%',
+  gap: '1.25rem',
+});
+
+export const PriceContent = css({
+  // step6
+  display: 'flex',
+  flex: 1,
+  flexDirection: 'column',
+  alignItems: 'flex-start',
+  alignSelf: 'stretch',
+  color: '100',
+  fontFamily: 'Pretendard',
+  fontSize: '1rem',
+  fontStyle: 'normal',
+  fontWeight: '500',
+  lineHeight: '1.4',
+  letterSpacing: '-0.04rem',
+});
+
+export const CanDeal = css({
+  // step6
+  display: 'flex',
+  justifyContent: 'center',
+  alignItems: 'center',
+  gap: '0.62rem',
+  color: '80',
+  fontFamily: 'Pretendard',
+  fontSize: '0.875rem',
+  fontStyle: 'normal',
+  fontWeight: '500',
+  lineHeight: '1.4',
+  letterSpacing: '-0.035rem',
+});
